Use the correct store for opening the recurrence blob
When a user edits an appointment as a delegate and the recurrenceblob is
bigger then 256K the wrong store is used in mapi_getprops as we always
pass the default store (logged in users store) and not the store of the
owner. As the correct store is already available in the
sendMeetingRequest function use this to resolve the problem.

References: KW-3600