Remove special HTML processing for Outlook styles and automatic linkify
This change removes the cleanup of certain Outlook specific styles. The
removed function was not efficient and can hang the entire browser if
the HTML it processes is bad, large or deeply nested.

The function `scanDOMForLinks` is inefficient and makes the browser/webapp
hang when the mail is complex. This change removes the implementation means
for the time being that feature is gone. Can be added properly later if
required.
￼

Reference KW-3717